deployments/huggingface/requirements.txt
torch
torchvision
transformers
opencv-python
pillow
numpy
gradio